tf-mnist/webapp/requirements.txt
Flask==0.10.1
tensorflow
tensorflow-serving-api
python-mnist
Pillow
gunicorn
